+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v1.3.8
 - Fix redis malfunction
 - Change bilibili-fanart logic
 - Add youtube changer status in user handler
+- Remove twitter with Quoted and Reply
 
 v1.3.7
 - Change backend to micro service